refactor client tests
